% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startdir.R
\name{startdir}
\alias{startdir}
\title{Generate apropriate directory structure inside of project folder.}
\usage{
startdir(type = c("package", "analysis"))
}
\arguments{
\item{"package"}{or "analysis".}
}
\value{
The apropriate directory structure eather \code{"package"} or \code{"analysis"}.
}
\description{
Generate apropriate directory structure inside of project folder.
}
\examples{
startdir("package")
}

