% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undir.R
\name{undir}
\alias{undir}
\title{Delete directory structure inside of project folder, which was generated by startdir function.}
\usage{
undir(type = c("package", "analysis"))
}
\arguments{
\item{"package"}{or "analysis".}
}
\value{
delete directory structure eather in format \code{"package"} or \code{"analysis"}.
}
\description{
Delete directory structure inside of project folder, which was generated by startdir function.
}
\examples{
undir("package")
}

